+++ virtool_cli/update/update.py
@@ -61,28 +61,4 @@ async def request_new_records(
         return []
 
     return record_data
-#
-#
-# async def fetch_upstream_accessions(
-#     taxid: int, included: list, excluded: list, logger: BoundLogger = base_logger
-# ) -> list:
-#     """
-#     Requests a list of all uninspected accessions associated with an OTU's taxon ID
-#
-#     :param listing: Corresponding catalog listing for this OTU
-#     :param logger: Optional entry point for an existing BoundLogger
-#     :return: A list of accessions from NCBI Genbank for the taxon ID,
-#         sans included and excluded accessions
-#     """
-#     logger = logger.bind(taxid=taxid)
-#     included_set = set(included)
-#
-#     try:
-#         upstream_accessions = await request_linked_accessions(taxon_id=taxid)
-#     except HTTPError:
-#         logger.error("Could not retrieve new accessions from NCBI.")
-#         return []
-#
-#     upstream_set = set(upstream_accessions)
-#
-#     return list(upstream_set.difference(included_set))
+
